Move tests for LowerStringField and LowerEmailField to tests.mongo.fields.test_basic
